add queries file for gathering all queries/mutations for organization
