+++ client/js/components/CreateComment/CreateComment.js
@@ -1,18 +1,11 @@
 import React from 'react';
 import {View, Text, TextInput, TouchableOpacity, Image} from 'react-native';
 import {Mutation} from '@apollo/react-components';
-import {gql} from 'apollo-boost';
-import styles from './styles';
-import {GET_ALL_POSTS} from '../../screens/Community/CommunityContainer';
 
-const CREATE_COMMENT = gql`
-  mutation createComment($authorId: ID, $postId: ID, $content: String!) {
-    createComment(authorId: $authorId, postId: $postId, content: $content) {
-      id
-      content
-    }
-  }
-`;
+import styles from './styles';
+import {GET_ALL_POSTS} from '../../config/apollo/queries';
+import {CREATE_COMMENT} from '../../config/apollo/queries'
+
 
 const CreateComment = ({postId, toggleCommentDisplay, viewer}) => {
   const [text, onChangeText] = React.useState();
