+++ client/js/screens/Community/CommunityContainer.js
@@ -1,48 +1,12 @@
 import React, {Component} from 'react';
 import {Text, TouchableOpacity, Image} from 'react-native';
 import Community from './Community';
-import {gql} from 'apollo-boost';
 import {Query} from '@apollo/react-components';
 import CircularLoader from '../../components/CircularLoader';
 import FavesContext from '../../context/FavesContext';
 import styles from './styles';
 import {withNavigation} from 'react-navigation';
-
-export const GET_ALL_POSTS = gql`
-  query {
-    allPosts(orderBy: createdAt_DESC) {
-      author {
-        id
-        firstName
-        lastName
-        photo {
-          url
-        }
-      }
-      type
-      id
-      createdAt
-      content
-      likes {
-        id
-      }
-      comments {
-        id
-        author {
-          firstName
-          lastName
-          photo {
-            url
-          }
-        }
-        content
-        likes {
-          id
-        }
-      }
-    }
-  }
-`;
+import {GET_ALL_POSTS} from '../../config/apollo/queries';
 
 class CommunityContainer extends Component {
   constructor(props) {
