+++ client/js/screens/Favourites/FavouritesContainer.js
@@ -1,47 +1,12 @@
 import React, {Component} from 'react';
-import {Text, TouchableOpacity, Image} from 'react-native';
+import {Text} from 'react-native';
 import Favourites from './Favourites';
-import {gql} from 'apollo-boost';
 import {Query} from '@apollo/react-components';
 import CircularLoader from '../../components/CircularLoader';
 import FavesContext from '../../context/FavesContext';
 import styles from './styles';
 import {withNavigation} from 'react-navigation';
-
-export const GET_ALL_POSTS = gql`
-  query {
-    allPosts(orderBy: createdAt_DESC) {
-      author {
-        firstName
-        lastName
-        photo {
-          url
-        }
-      }
-      type
-      id
-      createdAt
-      content
-      likes {
-        id
-      }
-      comments {
-        id
-        author {
-          firstName
-          lastName
-          photo {
-            url
-          }
-        }
-        content
-        likes {
-          id
-        }
-      }
-    }
-  }
-`;
+import {GET_ALL_POSTS} from '../../config/apollo/queries';
 
 class FavouritesContainer extends Component {
   static navigationOptions = {
